Query Scryfall Card Search API

Request card-level information. This API call operates in a similar manner to a manual Scryfall search.

scry_cards(query, .unique = "cards", .order = "name",
  direction = "auto", include_extras = FALSE,
  include_multilingual = FALSE, include_variations = FALSE,
  include_ids = FALSE, include_uris = FALSE, delay = 75)

Arguments

query
search query to pass to the API. A character string that looks the same as a search you'd manually enter into Scryfall. (see: https://scryfall.com/docs/syntax for docuementation on searching Scryfall)

.unique
How Scryfall handles cases where different versions of the same card match the query. "cards" (default) returns only one instance of card, "art" returns each instance with a different art "prints" returns all of them.

.order
How Scryfall sorts returned cards. "name" (default): card name, "set": set code and collector number, "released": release date, "rarity": rarity, "color": color, "usd": price in US dollars, "tix"; price in tickets on MTGO, "eur": price in Euros, "cmc": converted mana cost, "power": power, "toughenss": toughness, "edhrec": EDHREC rating, or "artist": artist name.

direction
Which direction cards are sorted in based on .order. "auto" (default): order that makes most sense for .order, "asc": ascending order, and "desc": descending order. "auto" sorts in descending order for prices ("usd", "tix", and "eur"), "release date", and "rarity". For all other .order options, "auto sorts in ascending order.

include_extras
Should results include extras like tokens or schemes.

include_multilingual
Should results include cards in all supported languages.

include_variations
Should results include rare case variants.

include_ids
Should results include ID variables (e.g. Oracle ID)

include_uris
Should results include URI variables (e.g. the URL for the card on Scryfall)

delay
Number of milliseconds scryr should wait between requests. (Scryfall asks for 50-100)

Value

a tibble of cards matching the search parameters

Examples

scry_cards("set:ktk")

#> # A tibble: 254 x 56
#>    object multiverse_ids name  lang  released_at layout highres_image mana_cost
#>    <chr>  <list>         <chr> <chr> <chr>       <chr>  <lgl>         <chr>    
#>  1 card   <int [1]>      Abom~ en    2014-09-26  normal TRUE          {3}{B}{G~
#>  2 card   <int [1]>      Abza~ en    2014-09-26  normal TRUE          {W}{B}{G}
#>  3 card   <int [1]>      Abza~ en    2014-09-26  normal TRUE          {3}      
#>  4 card   <int [1]>      Abza~ en    2014-09-26  normal TRUE          {3}{W}   
#>  5 card   <int [1]>      Abza~ en    2014-09-26  normal TRUE          {W}{B}{G}
#>  6 card   <int [1]>      Abza~ en    2014-09-26  normal TRUE          {2}{W}   
#>  7 card   <int [1]>      Abza~ en    2014-09-26  normal TRUE          {3}{W}{B~
#>  8 card   <int [1]>      Act ~ en    2014-09-26  normal TRUE          {2}{R}   
#>  9 card   <int [1]>      Aino~ en    2014-09-26  normal TRUE          {1}{W}   
#> 10 card   <int [1]>      Aino~ en    2014-09-26  normal TRUE          {5}{R}   
#> # ... with 244 more rows, and 48 more variables: cmc <dbl>, type_line <chr>,
#> #   oracle_text <chr>, power <chr>, toughness <chr>, colors <list>,
#> #   color_identity <list>, games <list>, reserved <lgl>, foil <lgl>,
#> #   nonfoil <lgl>, oversized <lgl>, promo <lgl>, reprint <lgl>,
#> #   variation <lgl>, set <chr>, set_name <chr>, set_type <chr>,
#> #   collector_number <chr>, digital <lgl>, rarity <chr>, watermark <chr>,
#> #   artist <chr>, border_color <chr>, frame <chr>, full_art <lgl>,
#> #   textless <lgl>, booster <lgl>, story_spotlight <lgl>, edhrec_rank <int>,
#> #   all_parts <list>, flavor_text <chr>, legalities.standard <chr>,
#> #   legalities.future <chr>, legalities.frontier <chr>,
#> #   legalities.modern <chr>, legalities.legacy <chr>, legalities.pauper <chr>,
#> #   legalities.vintage <chr>, legalities.penny <chr>,
#> #   legalities.commander <chr>, legalities.duel <chr>,
#> #   legalities.oldschool <chr>, prices.usd <chr>, prices.usd_foil <chr>,
#> #   prices.eur <chr>, prices.tix <chr>, loyalty <chr>

scry_cards("set:ktk wm:mardu")

#> # A tibble: 29 x 55
#>    object multiverse_ids name  lang  released_at layout highres_image mana_cost
#>    <chr>  <list>         <chr> <chr> <chr>       <chr>  <lgl>         <chr>    
#>  1 card   <int [1]>      Ankl~ en    2014-09-26  normal TRUE          {2}{R}{W~
#>  2 card   <int [1]>      Arro~ en    2014-09-26  normal TRUE          {3}{R}{R}
#>  3 card   <int [1]>      Bell~ en    2014-09-26  normal TRUE          {3}{B}   
#>  4 card   <int [1]>      Bloo~ en    2014-09-26  normal TRUE          {B}      
#>  5 card   <int [1]>      Butc~ en    2014-09-26  normal TRUE          {1}{R}{W~
#>  6 card   <int [1]>      Cany~ en    2014-09-26  normal TRUE          {4}{R}   
#>  7 card   <int [1]>      Chie~ en    2014-09-26  normal TRUE          {W}{B}   
#>  8 card   <int [1]>      Crac~ en    2014-09-26  normal TRUE          {R}{W}{B}
#>  9 card   <int [1]>      Fire~ en    2014-09-26  normal TRUE          {W}      
#> 10 card   <int [1]>      Hord~ en    2014-09-26  normal TRUE          {1}{R}   
#> # ... with 19 more rows, and 47 more variables: cmc <dbl>, type_line <chr>,
#> #   oracle_text <chr>, power <chr>, toughness <chr>, colors <list>,
#> #   color_identity <list>, games <list>, reserved <lgl>, foil <lgl>,
#> #   nonfoil <lgl>, oversized <lgl>, promo <lgl>, reprint <lgl>,
#> #   variation <lgl>, set <chr>, set_name <chr>, set_type <chr>,
#> #   collector_number <chr>, digital <lgl>, rarity <chr>, watermark <chr>,
#> #   flavor_text <chr>, artist <chr>, border_color <chr>, frame <chr>,
#> #   full_art <lgl>, textless <lgl>, booster <lgl>, story_spotlight <lgl>,
#> #   edhrec_rank <int>, all_parts <list>, legalities.standard <chr>,
#> #   legalities.future <chr>, legalities.frontier <chr>,
#> #   legalities.modern <chr>, legalities.legacy <chr>, legalities.pauper <chr>,
#> #   legalities.vintage <chr>, legalities.penny <chr>,
#> #   legalities.commander <chr>, legalities.duel <chr>,
#> #   legalities.oldschool <chr>, prices.usd <chr>, prices.usd_foil <chr>,
#> #   prices.eur <chr>, prices.tix <chr>